@@ -1,15 +1,21 @@
-module.exports = {
+import invisibleRecaptcha from './invisible-recaptcha.vue';
+
+export default {
 
     data () {
         return  _.merge({
             status: 0,
             form: null,
-            formData: new FormData
+            formData: new FormData,
         }, this.$options.form);
     },
 
-
     methods: {
+
+        submit () {
+            if (this.$pagekit.recaptcha) this.$refs.recaptcha.execute();
+            else this.save();
+        },
 
         save () {
 
@@ -22,8 +28,6 @@ module.exports = {
             this.status = 1;
 
             this.$http.post('api/form', this.formData).then(function (res) {
-
-                console.log(res);
 
                 this.$notify('Success.');
 
@@ -41,12 +45,27 @@ module.exports = {
             });
         },
 
-        components: {
-            loader: {
-                template:  `<svg class="pk-loader" width="30" height="30" viewBox="0 0 30 30" xmlns="http://www.w3.org/2000/svg">
-                                <g><circle cx="0" cy="0" r="13" fill="none" stroke-width="1"/></g>
-                            </svg>`
-            }
+        onCaptchaVerified (recaptchaToken) {
+            this.$refs.recaptcha.reset();
+            this.formData.append('recaptcha', recaptchaToken);
+            this.save();
+        },
+
+        onCaptchaExpired () {
+            this.$refs.recaptcha.reset();
+        },
+
+        onCaptchaError (error) {
+            this.$notify(error);
+        },
+
+        render(id)
+        {
+            console.log(id);
         }
+    },
+
+    components: {
+        invisibleRecaptcha
     }
 }
