@@ -1,6 +1,10 @@
 <?php
 
 use SAB\Form\Plugin\FormPlugin;
+use Pagekit\View\Event\ViewEvent;
+use Pagekit\View\View;
+use Pagekit\View\Asset\AssetManager;
+use Pagekit\Event\Event;
 
 return [
 
@@ -29,14 +33,33 @@ return [
 
     ],
 
+    'config' => [
+        'recaptcha' => [
+            'sitekey' => '',
+            'secret' => ''
+        ]
+    ],
+
     'events' => [
 
-        'view.scripts' => function ($event, $scripts) {
-            $scripts->register('forms', 'form:app/bundle/forms.js', ['vue']);
+        'view.init' => function (Event $event, View $view) use ($app) {
+            $view->data('$pagekit', [
+                'recaptcha' => $app['config']->get('form')->get('recaptcha.sitekey')
+            ]);
         },
 
-        'view.styles' => function ($event, $styles) {
+        'view.scripts' => function (Event $event, AssetManager $scripts) {
+            $scripts->register('g-recaptcha', 'https://www.google.com/recaptcha/api.js');
+            $scripts->register('forms', 'form:app/bundle/forms.js', ['vue', 'g-recaptcha']);
+        },
+
+        'view.styles' => function (Event $event, AssetManager $styles) {
             $styles->register('form', 'form:css/form.css');
+        },
+
+        'view.system/site/admin/settings' => function (ViewEvent $event, View $view) use ($app) {
+            $view->script('site-recaptcha', 'form:app/bundle/site-recaptcha.js', 'site-settings');
+            $view->data('$form', $this->config());
         }
 
     ]
